Use of Expression tags <%= statement %>
Also, looking into the jasper component who compiles the jsp file to the servlet and put code inside the service method.

Note: Do not end your statement with semicolon in case of expression tag.

It is mainly used to print the values of variable or method.

The code placed within JSP expression tag is written to the output stream of the response.
So you need not write out.print() to write data.